Refactored and removed unneeded code (lots)
Added comments as to why I deleted the stuff. I figured that it'd be better to build every single feature up from scratch so we have experience with it all and we can explain all the code. This is mainly just my opinion and since this is a pull request, it doesn't have to be accepted by the group. Hopefully the group sees the logic in this, though.